+++ packages/driver/src/cypress.ts
@@ -137,6 +137,7 @@ class $Cypress {
   originalConfig: any
   config: any
   env: any
+  expose: any
   getTestRetries: any
   Cookies!: ICookies
   ProxyLogging: any
@@ -254,7 +255,7 @@ class $Cypress {
     // not we're in a text terminal, but we keep this
     // as a separate property so we can potentially
     // slice up the behavior
-    config.isInteractive = !config.isTextTerminal || config.env.INTERNAL_SIMULATE_OPEN_MODE
+    config.isInteractive = !config.isTextTerminal || config.env?.INTERNAL_SIMULATE_OPEN_MODE
 
     // true if this Cypress belongs to a cross origin spec bridge
     this.isCrossOriginSpecBridge = config.isCrossOriginSpecBridge || false
@@ -271,9 +272,9 @@ class $Cypress {
 
     // TODO: env is unintentionally preserved between soft reruns unlike config.
     // change this in the NEXT_BREAKING
-    const { env } = config
+    const { env, expose } = config
 
-    config = _.omit(config, 'env', 'rawJson', 'remote', 'resolved', 'scaffoldedFiles', 'state', 'testingType', 'isCrossOriginSpecBridge')
+    config = _.omit(config, 'env', 'expose', 'rawJson', 'remote', 'resolved', 'scaffoldedFiles', 'state', 'testingType', 'isCrossOriginSpecBridge')
 
     _.extend(this, browserInfo(config))
 
@@ -311,7 +312,18 @@ class $Cypress {
       return validateConfig(this.state, config, skipConfigOverrideValidation)
     })
 
-    this.env = $SetterGetter.create(env)
+    const isAllowCypressEnvEnabled = this.config('allowCypressEnv')
+
+    const failCypressEnvWithWarning = (key) => {
+      const err = $errUtils.errByPath('config.allow_cypress_env', { key })
+
+      $utils.warning(err.message)
+
+      throw err
+    }
+
+    this.env = !isAllowCypressEnvEnabled ? failCypressEnvWithWarning : $SetterGetter.create(env)
+    this.expose = $SetterGetter.create(expose)
     this.getTestRetries = function () {
       const testRetries = this.config('retries')