+++ packages/reporter/src/header/OpenFileInIDEButton.tsx
@@ -1,7 +1,7 @@
 import Button from '@cypress-design/react-button'
 import React from 'react'
 import events from '../lib/events'
-import { IconWindowCodeEditor } from '@cypress-design/react-icon'
+import { IconWindowCodeEditorSmall } from '@cypress-design/react-icon'
 import { FileDetails } from '@packages/types'
 import cx from 'classnames'
 
@@ -12,7 +12,7 @@ interface Props {
 
 export const OpenFileInIDEButton = ({ fileDetails, className }: Props) => {
   return (<>
-    <Button size='20' variant='outline-dark' className={cx('open-in-ide-button', className)} onClick={() => events.emit('open:file:unified', fileDetails)}><IconWindowCodeEditor strokeColor='gray-500' fillColor='gray-900' /> Open in IDE </Button>
+    <Button size='20' variant='outline-dark' className={cx('open-in-ide-button', className)} onClick={() => events.emit('open:file:unified', fileDetails)}><IconWindowCodeEditorSmall strokeColor='gray-500' fillColor='gray-900' /> Open in IDE </Button>
     <span className={cx('button-hover-shadow')} />
   </>
   )