+++ system-tests/project-fixtures/react/cypress-webpack-dev-server-async-config.config.ts
@@ -3,6 +3,7 @@ import * as path from 'path'
 import * as fs from 'fs'
 
 export default defineConfig({
+  allowCypressEnv: false,
   component: {
     devServer: {
       framework: 'react',