+++ system-tests/projects/plugin-code-coverage/cypress.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+  allowCypressEnv: false,
   projectId: 'abc123',
   experimentalInteractiveRunEvents: true,
   e2e: {