+++ system-tests/projects/studio/cypress.config.js
@@ -1,3 +1,4 @@
 module.exports = {
+  'allowCypressEnv': false,
   'experimentalStudio': true,
 }