+++ system-tests/projects/ts-proj-with-module-esnext/cypress.config.ts
@@ -1,6 +1,7 @@
 import { asyncGreeting } from './greeting'
 
 module.exports = {
+  'allowCypressEnv': false,
   'supportFolder': false,
   'e2e': {
     'supportFile': false,